Refactor: move steps & remove getValues
