+++ src/components/Auth/Signup/Step/Step.module.scss
